requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.11
djangorestframework==3.15.1
kombu==5.3.7
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
